requirements.txt
asgiref==3.8.1  
backports.zoneinfo==0.2.1  
Django==4.2.13 
django-rest-framework==0.1.0  
djangorestframework==3.15.2 
pip==20.0.2 
psycopg2==2.9.9  
psycopg2-binary==2.9.9  
setuptools==44.0.0 
sqlparse==0.5.0  
typing-extensions==4.12.2 
tzlocal==5.2    
wheel==0.43.0 
